Created schema validation for FEE configuration

+++ DataFormats/Detectors/FIT/FT0/include/DataFormatsFT0/FeeConfiguration.h
@@ -18,8 +18,8 @@
 namespace o2::ft0
 {
 struct TriggersConfig {
-  float vertexTimeLowThreshold{o2::fit::config_helpers::DefaultValue};
-  float vertexTimeHighThreshold{o2::fit::config_helpers::DefaultValue};
+  int16_t vertexTimeLowThreshold{o2::fit::config_helpers::DefaultValue};
+  int16_t vertexTimeHighThreshold{o2::fit::config_helpers::DefaultValue};
   uint16_t semicentralA{o2::fit::config_helpers::DefaultValue};
   uint16_t semicentralC{o2::fit::config_helpers::DefaultValue};
   uint16_t centralA{o2::fit::config_helpers::DefaultValue};
